Fix "desktop" cloud deploy
If you fork the repo you can create a complete deployment in your forked repo, but if you just want to test deployment to the cloud
the "desktop" deploy to cloud just deploys a single stage.

This update fixes that setup and support for attach/detach.

related to PolicyEngine/issues#385